Fixing a couple of issues: - sometimes the session was not restored when using one-shot - scriptname in /var/log/syslog only (removed the path) - being more verbose in syslog output (now every action will be shown) - replaced the useless dconf read for one-shot by empty dir check instead - sometimes dex would stop when there was a failure when using the autostart   option for a directory. so I switched over to not use the autostart but   executing every single desktop file instead

+++ restore-session.sh
@@ -15,6 +15,10 @@
 # 4) Open startup programs settings tool of Gnome and add the full path to this script
 # 5) logout and enjoy your next login
 #
+# Force to re-read a saved gnome session or if you use the one-shot and want to change it:
+# a) do step 2 + 3
+# b) do: "rm -R ~/.config/gnome-session/dex/"
+# c) do step 5
 ########################################################################################
 
 # REPLACE THE FOLLOWING TO YOUR COPY OF DEX:
@@ -22,23 +26,34 @@ DEX=~/install/scripts/dex.py
 
 ######### C O D E ##################################################
 
+ME="${0##*/}"
 GNOMEDIR=~/.config/gnome-session/saved-session
 EXECDIR=~/.config/gnome-session/dex
 
-logger -t "$0" "Gnome session restored started."
+logger -t "$ME" "Gnome session restored started."
 # check if we have a valid session dir
 [ ! -d $GNOMEDIR ]&& logger -t "$0" "ABORTED execution because $GNOMEDIR does not exist. Do you have saved a Gnome session?" && exit
 
-# setup env
-rm -rf $EXECDIR
-mkdir -p $EXECDIR
+RESAVE=$(dconf read /org/gnome/gnome-session/auto-save-session)
+DEXTHERE=$()
 
-# take gnome saved session and copy it to working dir
-find "$GNOMEDIR" -type f -name "*.desktop" -exec cp {} "$EXECDIR" \;
-# made them executable so remove gnome junk
-find "$EXECDIR" -type f -name "*.desktop" -exec sed -i 's/--sm-c.*//g' {} \; 
+# when the one-shot is true or when the DEX dir does not exists enforce the session copy
+if [ ! -d $EXECDIR ]||[ ! "$(ls -A $EXECDIR)" ]||[ x$RESAVE == "xtrue" ];then
+        logger -t "$ME" "Copy and prepare previous session.."
+	# setup env
+	[ -d $EXECDIR ] && rm -rf $EXECDIR
+	mkdir -p $EXECDIR
+	# take gnome saved session and copy it to working dir
+	find "$GNOMEDIR" -type f -name "*.desktop" -exec cp {} "$EXECDIR" \;
+	# made them executable so remove gnome junk
+	find "$EXECDIR" -type f -name "*.desktop" -exec sed -i 's/--sm-c.*//g' {} \; 
+	find "$EXECDIR" -type f -name "*.desktop" -exec sed -i 's/--session.*//g' {} \; 
+else
+	logger -t "$ME" "Using already prepared previous session.."
+fi
 
 # restore the session
-$DEX -a -s "$EXECDIR"
+logger -t "$ME" "$(find "$EXECDIR" -type f -name "*.desktop" -exec $DEX -v {} \;)"
+#$DEX -a -s "$EXECDIR"
 
-logger -t "$0" "Gnome session restored ended."
+logger -t "$ME" "Gnome session restored ended."
